Ported S2PolygonBuilder support for edge splicing from the C++ S2 library. Edge splicing allows edges to be considered coincident even if they do not share vertices, splicing together lines in a T formation, etc.
Ported new S2PolygonBuilderTest tests from the C++ S2 library. The tests address new features like edge splicing, but also provide significantly better coverage of the preexisting S2PolygonBuilder features.

The S2PolygonBuilderTest class now disables logging from S2Polygon and S2Loop since it was 100% spurious and interfered with the important logs coming from S2PolygonBuilder.

Ported the startingVertices approach to consistent vertex ordering from C++ S2PolygonBuilder to Java. This should allow the builder to produce loops with the same order as the C++ library, except in rare cases where the differing numerical precisions of the C++ and Java libraries results in differing topologies.

Changed S2PolygonBuilder.Options from an enum to a static inner class, with 4 static final constant values of that class that allow all prior references to the enums to work the same way as before, even though the reference is now a proper class. Prior to this change, setting the vertex merge distance affected all other code in the JVM using the same Options enum. After this change, independent Options instances may now be created to hold custom values that cannot impact other code. In the future it may be desirable to make Options immutable, but that will affect a few remote code locations and is outside the scope of this CL; for the moment, it is enough that this change allows the S2 tests to succeed regardless of the order they are called in.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=29566596