Make the geometry library gwt-compatible
There are a couple StrictMath methods that aren't available in GWT, so for now
they just throw an exception.  I'll fix them in a follow-on CL.  The tests that
fail as a result of this I have disabled for now.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=47111174