Adds support for polygons larger than a hemisphere. Before this CL, a clockwise (right-turning) loop was invalid, and you could call S2Loop.normalize() to fix validity issues.  Now, a right-turning loop is valid: interior area is still to the left, so e.g. a small right-turning triangle is the complement of the same edges: it encloses the whole sphere, except that small area.  This is required to allow very large features, like a whole-world water bodies mask.
This CL also adds support for full and empty polygons.  So the union of two complementary hemispheres is the full polygon, the full polygon minus the empty polygon is the full polygon, the intersection between empty and anything is empty, etc.

This CL borrows heavily from the C++ implementation of the same concept.  In particular, the internal representation of the full and empty loops as having a single vertex in the south or north hemisphere has been maintained, for eventual serialization compatibility.

This CL also basically re-ports the relational tests from s2polygon_test.cc, which had drifted out of alignment with those in C++.  To get the tests passing, it was necessary to move S2.origin() near the north poled and add some robust bounds checking fixes.  There are also some cherry picked single-line test changes.

Finally, this CL includes S2Shape, which won't strictly be needed until the S2ShapeIndex goes in.  However, some of the new algorithms that refer to S2Shape methods may do so if I port it now, and implementing those methods is trivial, so I have frontloaded that small work to this CL.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=61276472