A few S2Shape and S2ShapeIndex API changes:
- S2ClippedShape method "int shapeId()" is replaced by new public method "S2Shape shape()". This is a cleaner public API that relegates the int shapeId to an internal detail of index construction. By avoiding lookups, indexing is ~3% faster, queries are ~10% faster, the index is the same size with compressed pointers, and without compressed pointers the index is only ~10% larger.

- S2Shape.MutableEdge is now a final class instead of an interface, so clients don't have to implement their own MutableEdge, and since we ended up requiring only one implementation.

Also switched the internal Cell implementations to store the containsCenter boolean *on the class*, since it can only have two values. This mucks up the internal details of creating Cells a little bit, but avoids a padding penalty that makes the index ~30% larger in some cases. We were XORing this bit with the int shapeId but no longer have that option, so we store on the class instead.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=72983010