Fix incorrect uses of hashCode on arrays.
When computing a hashcode for an array, typically you want a hashcode that
depends on the value of each element in the array.  However, many of the common
hashcode methods actually return an identity hashcode rather than one that
depends on the contents.  This CL flags these identity hashcodes on arrays, and
changes them to a hashcode that does depend on the array contents.

We will be making this an error in the Java compiler shortly, so it is
necessary to fix existing occurrences first to avoid breaking builds.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=86969996